test: split authentication factories
